Moves resources to separate package
